Fixed UV PIN and rain sensor pin

+++ main/Data_Capture.h
@@ -21,7 +21,7 @@ float gIrradiance;
 float gPAR;
 
 /* Uv Sensor */
-byte gUvindex;
+float gUvindex;
 
 /* Wind direction */
 int gWinddirectiondata;
